+++ pch.h
@@ -1,5 +1,5 @@
 
-/** $VER: pch.h (2025.09.15) P. Stuer **/
+/** $VER: pch.h (2025.09.22) P. Stuer **/
 
 #pragma once
 
@@ -48,32 +48,5 @@
 EXTERN_C IMAGE_DOS_HEADER __ImageBase;
 #define THIS_HINSTANCE ((HINSTANCE) &__ImageBase)
 #endif
-
-/// <summary>
-/// A more sane way of representing a rectangle
-/// </summary>
-struct rect_t
-{
-    rect_t & operator = (const D2D1_RECT_F & other) noexcept
-    {
-        *this = other;
-
-        return *this;
-    }
-
-    operator D2D1_RECT_F () const noexcept
-    {
-        return { x1, y1, x2, y2 };
-    }
-
-    D2D1_SIZE_F Size() const noexcept { return { std::abs(x1 - x2), std::abs(y1 - y2) }; }
-    FLOAT Width() const noexcept { return std::abs(x2 - x1); }
-    FLOAT Height() const noexcept { return std::abs(y2 - y1); }
-
-    FLOAT x1;
-    FLOAT y1;
-    FLOAT x2;
-    FLOAT y2;
-};
 
  constexpr WCHAR NegativeInfinity[3] = { '-', 0x221E, '\0' };
